frameworks/fm_google_adk/pyproject.toml
[project]
name = "google-adk-bbh-benchmark"
version = "0.1.0"
description = "Google ADK framework implementation for BBH benchmarking"
requires-python = ">=3.11"
dependencies = [
    "openai",
    "datasets",
    "google-adk>=1.12.0",
    "litellm>=1.76.3",
]
setup = "ready"

[tool.uv]
managed = true
